fix: resolve test warnings and rewrite network page tests for new architecture
