+++ packages/dashboard/src/i18n/locales/ar.json
@@ -214,6 +214,9 @@
       "teams": "الفرق",
       "storage": "إجمالي التخزين"
     },
+    "overview": {
+      "networkInfo": "معلومات الشبكة"
+    },
     "identity": {
       "title": "هوية الشبكة",
       "description": "قم بتكوين العلامة التجارية وهوية شبكتك.",
